% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{boot_hr}
\alias{boot_hr}
\title{Bootstrap hazards ratios of two survival models
\code{boot_hr} computes bootstrap hazard ratios (HR) of two survival models (model 1 vs. 2) or one survival model with multiple treatments (rx 1 vs. 2)}
\usage{
boot_hr(
  surv_model1 = NULL,
  surv_model2 = NULL,
  rx1 = NULL,
  rx2 = NULL,
  rx = F,
  surv_model_rx = NULL,
  times,
  B = 100
)
}
\arguments{
\item{surv_model1}{first survival model.}

\item{surv_model2}{second survival model.}

\item{rx1}{first treatment arm (if rx = T).}

\item{rx2}{second treatment arm (if rx = T).}

\item{rx}{whether to model treatment arms within one model.
Default = FALSE.}

\item{surv_model_rx}{survival model (if rx = T), can only choose weibull or gamma models.}

\item{times}{time horizon the extrapolation of the survival model is done over.}

\item{B}{number of bootstrap samples.}
}
\value{
dataframe of hazard ratio statistics (2.5\% percentile, median, 97.5\% percentile, time points)
}
\description{
Bootstrap hazards ratios of two survival models
\code{boot_hr} computes bootstrap hazard ratios (HR) of two survival models (model 1 vs. 2) or one survival model with multiple treatments (rx 1 vs. 2)
}
