% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{boot_muhaz}
\alias{boot_muhaz}
\title{Non-parametric hazard bootstrap
\code{boot_hr} computes non-parametric bootstrap hazards from time-to-event data.}
\usage{
boot_muhaz(surv_data, time, status, B)
}
\arguments{
\item{surv_data}{survival (time-to-event) data.}

\item{time}{name of time variable in survival data (character variable).}

\item{status}{name of status variable survival data (character variable).}

\item{B}{number of bootstrap samples.}
}
\value{
list of objects (time points, hazard, hazard CI, bootstrapped time points).
}
\description{
Non-parametric hazard bootstrap
\code{boot_hr} computes non-parametric bootstrap hazards from time-to-event data.
}
