% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calc_prevalence}
\alias{calc_prevalence}
\title{Calculate prevalence}
\usage{
calc_prevalence(l_m_M, v_names_sick_states, v_names_dead_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a dataframe containing prevalence of specified health states for each strategy
}
\description{
\code{plot_prevalence} calculate the prevalence for different health states.
}

