% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calc_sick}
\alias{calc_sick}
\title{Calculate state proportions}
\usage{
calc_sick(l_m_M, v_names_sick_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a dataframe containing proportions in specified states for each strategy
}
\description{
\code{calc_surv} calculates the proportions of the cohort in specified states
}
