% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calc_surv}
\alias{calc_surv}
\title{Calculate survival probabilities}
\usage{
calc_surv(l_m_M, v_names_death_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a dataframe containing survival probabilities for each strategy
}
\description{
\code{calc_surv} calculates the survival probabilities.
}
