% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list_elements.R
\name{check_list_elements}
\alias{check_list_elements}
\title{Check if the each item in the list contains information.}
\usage{
check_list_elements(l_list, err_stop = TRUE, verbose = TRUE)
}
\arguments{
\item{l_list}{A list with parameter values.}

\item{err_stop}{Logical variable to stop model run if set up as TRUE.
Default = TRUE.}

\item{verbose}{Logical variable to indicate print out of messages.
Default = TRUE}
}
\value{
Information about the validity of the list
}
\description{
\code{check_list_elements} checks if item in the list contains a value
}
