% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{check_PFS_OS}
\alias{check_PFS_OS}
\title{Print a warning message if PFS > OS}
\usage{
check_PFS_OS(Sick)
}
\arguments{
\item{Sick}{vector (or matrix) of PFS - OS probabilities}
}
\value{
a warning message if PFS > OS
}
\description{
\code{check_PFS_OS} prints a warning message if PFS > OS.
}

