% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{create_at_risk_table}
\alias{create_at_risk_table}
\title{Function to create at-risk table.}
\usage{
create_at_risk_table(survival_time, Years, AtRisk)
}
\arguments{
\item{survival_time}{vector of survival times.}

\item{Years}{vector of time intervals.}

\item{AtRisk}{vector of number of patients at risk.}
}
\value{
at-risk table (dataframe).
}
\description{
\code{create_at_risk_table} creates at-risk table.
}

