% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{cumhaz_to_haz}
\alias{cumhaz_to_haz}
\title{Converts cumulative hazards to hazard rates}
\usage{
cumhaz_to_haz(cumhaz)
}
\arguments{
\item{cumhaz}{vector of cumulative hazards}
}
\value{
vector of hazard rates
}
\description{
\code{cumhaz_to_haz} converts cumulative hazards to hazard rates across time points.
}
