% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{expected_surv}
\alias{expected_surv}
\title{Calculate expected survival.}
\usage{
expected_surv(time, surv)
}
\arguments{
\item{time}{vector of time to estimate probabilities.}

\item{surv}{vector of survival probabilities.}
}
\value{
expected survival.
}
\description{
\code{expected_surv} calculates expected survival (area under survival curve).
}
