% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{find_interval_limits}
\alias{find_interval_limits}
\title{Determine which rows the upper and lower values of each interval are (in the survival data set).}
\usage{
find_interval_limits(start_time, surv_time)
}
\arguments{
\item{start_time}{survival times.}
}
\value{
matrix of limits.
}
\description{
\code{find_interval_limits} determines which rows the upper and lower values of each interval are (in the survival data set).
}

