% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fit.fun}
\alias{fit.fun}
\title{Fit multiple survival models survival data}
\usage{
fit.fun(time, status, data = data, extrapolate = FALSE, times, k = 2)
}
\arguments{
\item{time}{numeric vector of time to estimate probabilities.}

\item{status}{numeric vector of event status.}

\item{data}{dataframe containing the time and status variables.}

\item{extrapolate}{extrapolate beyond model time horizon.
Default = FALSE.}

\item{times}{time horizon the extrapolation is done over.}
}
\value{
a list containing all survival model objects.
}
\description{
\code{fit.fun} fits multiple survival models to survival data using survHE.
}
