% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fit.models.cure}
\alias{fit.models.cure}
\title{Fit parametric mixture cure survival models for health economic evaluations.}
\usage{
fit.models.cure(formula = NULL, data, distr = NULL, method = "mle", ...)
}
\arguments{
\item{formula}{a formula specifying the model to be used, in the form Surv(time,event)~treatment\link{+covariates} for flexsurv.}

\item{data}{A data frame containing the data to be used for the analysis. This must contain data for the 'event' variable. In case there is no censoring, then event is a column of 1s.}

\item{distr}{a (vector of) string(s) containing the name(s) of the model(s) to be fitted. Available options are: flexsurv: "exponential","gamma","genf","gengamma","gompertz","weibull", "weibullPH","loglogistic","lognormal" INLA: "exponential","weibull","lognormal","loglogistic" hmc: "exponential","gamma","genf","gengamma","gompertz","weibull","weibullPH", "loglogistic","lognormal".}
}
\value{
A model object.
}
\description{
\code{fit.models.cure} fits parametric mixture cure survival models for health economic evaluations.
}

