% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fit.mstate}
\alias{fit.mstate}
\title{Fit multi-state model.}
\usage{
fit.mstate(
  time,
  status,
  trans,
  data = data,
  add = FALSE,
  extrapolate = FALSE,
  times
)
}
\arguments{
\item{time}{numeric vector of time to estimate probabilities.}

\item{status}{numeric vector of event status.}

\item{trans}{matrix of transition history.}

\item{data}{dataframe containing the time and status variables.}

\item{add}{add superimposed curves to the KM plot.
Default = FALSE.}

\item{extrapolate}{extrapolate beyond model time horizon.
Default = FALSE.}

\item{times}{time horizon the extrapolation is done over.}
}
\value{
Multi-state model fit.
}
\description{
\code{fit.mstate} fits multi-state model.
}

