% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cea_table.R
\name{format_table_cea}
\alias{format_table_cea}
\title{Format CEA table}
\usage{
format_table_cea(table_cea)
}
\arguments{
\item{table_cea}{a dataframe object - table with CEA results}
}
\value{
a dataframe object - formatted CEA table
}
\description{
\code{format_table_cea} formats the CEA table.
}
