% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{gen_data}
\alias{gen_data}
\title{Generate data for partitioned survival models (PFS and OS data).}
\usage{
gen_data(n_pat, n_years)
}
\arguments{
\item{n_pat}{number of patients.}

\item{n_years}{follow-up period.}
}
\value{
generated survival data.
}
\description{
\code{gen_data} generates survival data for overall survival (OS) and progression-free survival (PFS).
}
