% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_cycle_corrections.R
\name{gen_wcc}
\alias{gen_wcc}
\title{Within-cycle correction (WCC)}
\usage{
gen_wcc(n_cycles, method = c("Simpson1/3", "half-cycle", "none"))
}
\arguments{
\item{n_cycles}{number of cycles}

\item{method}{The method to be used for within-cycle correction.}
}
\value{
A vector of length \code{n_cycles + 1} with within-cycle corrections
}
\description{
\code{gen_wcc} generates a vector of within-cycle corrections (WCC).
}
\details{
The default method is an implementation of Simpson's 1/3rd rule that
generates a vector with the first and last entry with 1/3 and the odd and
even entries with 4/3 and 2/3, respectively.

Method "\code{half-cycle}" is the half-cycle correction method that
generates a vector with the first and last entry with 1/2 and the rest equal
to 1.

Method "\code{none}" does not implement any within-cycle correction and
generates a vector with ones.
}
\examples{
# Number of cycles
n_cycles <- 10
gen_wcc(n_cycles = n_cycles, method = "Simpson1/3")
gen_wcc(n_cycles = n_cycles, method = "half-cycle")
gen_wcc(n_cycles = n_cycles, method = "none")

}
\references{
\enumerate{
\item Elbasha EH, Chhatwal J. Myths and misconceptions of within-cycle
correction: a guide for modelers and decision makers. Pharmacoeconomics.
2016;34(1):13-22.
\item Elbasha EH, Chhatwal J. Theoretical foundations and practical
applications of within-cycle correction methods. Med Decis Mak.
2016;36(1):115-131.
}
}
