% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{get_DARTH_cols}
\alias{get_DARTH_cols}
\title{Get DARTH colors}
\usage{
get_DARTH_cols()
}
\value{
a string containing DARTH color codes
}
\description{
\code{get_DARTH_cols} retrieves the color codes for DARTH colors.
}
