% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_names_states.R
\name{get_v_names_states}
\alias{get_v_names_states}
\title{Get currently registered health-state names (if any).}
\usage{
get_v_names_states()
}
\value{
Character vector or NULL if unset.
}
\description{
Get currently registered health-state names (if any).
}
