% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{model.dist}
\alias{model.dist}
\title{Calculate survival probabilities given a survival distribution and its parameter values.}
\usage{
model.dist(dist.v, d.data, t)
}
\arguments{
\item{dist.v}{a character string specifying the name of the survival distribution.}

\item{d.data}{a vector of parameter values of the survival distribution.}

\item{t}{a vector of time points to calculate the survival probabilities at.}
}
\value{
A vector of survival probabilities.
}
\description{
\code{model.dist} calculates survival probabilities given a survival distribution and its parameter values.
}
