% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{model.rmvnorm}
\alias{model.rmvnorm}
\title{Randomly draw parameter values of survival models from multivariate normal distribution.}
\usage{
model.rmvnorm(dist.v, d.data, vc.data, n_sim, seed = 421)

model.rmvnorm(dist.v, d.data, vc.data, n_sim, seed = 421)
}
\arguments{
\item{dist.v}{a character string specifying the name of the survival distribution.}

\item{d.data}{a vector of mean parameter estimates of the survival distribution.}

\item{vc.data}{variance-covariance matrix (a matrix) of parameter estimates of the survival distribution.}

\item{n_sim}{number of random samples to draw.
Default = 100.}

\item{seed}{seed for random number generation.
Default = 421.}
}
\value{
A matrix of drawn parameter values.

A matrix of drawn parameter values.
}
\description{
\code{model.rmvnorm} randomly draws parameter values of survival models from multivariate normal distribution.

\code{model.rmvnorm} randomly draws parameter values of survival models from multivariate normal distribution.
}
