% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_conversion.R
\name{odds_to_prob}
\alias{odds_to_prob}
\title{Convert a odds to a probability}
\usage{
odds_to_prob(odds)
}
\arguments{
\item{odds}{a scalar of vector of odds}
}
\value{
a scalar or vector of probabilities
}
\description{
\code{odds_to_prob} convert an odds to a probability.
}
