% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{partsurv}
\alias{partsurv}
\title{Fit partitioned survival model to survival data}
\usage{
partsurv(
  pfs_survHE = NULL,
  os_survHE = NULL,
  l_d.data = NULL,
  l_vc.data = NULL,
  par = FALSE,
  dist_name_PFS = NULL,
  dist_name_OS = NULL,
  choose_PFS = NULL,
  choose_OS = NULL,
  time = times,
  v_names_states,
  PA = FALSE,
  n_sim = 100,
  seed = 421
)
}
\arguments{
\item{pfs_survHE}{survHE obj fitting PFS.}

\item{os_survHE}{survHE obj fitting OS.}

\item{l_d.data}{list of mean parameter estimates (for PFS and OS)}

\item{l_vc.data}{list of variance-covariance matrices of parameter estimates (for PFS and OS)}

\item{par}{use parameter mean estimates and variance-covariance matrix instead of IPD
Default = FALSE}

\item{dist_name_PFS}{specify PFS distribution when par = T.
Choose from: Exponential, Weibull (AFT), Gamma, log-Normal, log-Logistic, Gompertz, Exponential Cure, Weibull (AFT) Cure, Gamma Cure, log-Normal Cure, log-Logistic Cure, Gompertz Cure.}

\item{dist_name_OS}{specify OS distribution when par = T.
Choose from: Exponential, Weibull (AFT), Gamma, log-Normal, log-Logistic, Gompertz, Exponential Cure, Weibull (AFT) Cure, Gamma Cure, log-Normal Cure, log-Logistic Cure, Gompertz Cure.}

\item{choose_PFS}{preferred PFS distribution.}

\item{choose_OS}{preferred OS distribution.}

\item{time}{numeric vector of time to estimate probabilities.}

\item{v_names_states}{vector of state names.}

\item{PA}{run probabilistic analysis.
Default = FALSE.}

\item{n_sim}{number of PA simulations.
Default = 100.}

\item{seed}{seed for random number generation.
Default = 421.}
}
\value{
a list containing Markov trace, expected survival, survival probabilities, transition probabilities.
}
\description{
\code{partsurv} fits partitioned survival model to survival data.
}
