% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.icers}
\alias{plot.icers}
\title{Plot of ICERs}
\usage{
\method{plot}{icers}(
  x,
  txtsize = 12,
  currency = "$",
  effect_units = "QALYs",
  label = c("frontier", "all", "none"),
  label_max_char = NULL,
  plot_frontier_only = FALSE,
  alpha = 1,
  n_x_ticks = 6,
  n_y_ticks = 6,
  xbreaks = NULL,
  ybreaks = NULL,
  xlim = NULL,
  ylim = NULL,
  xexpand = expansion(0.1),
  yexpand = expansion(0.1),
  max.iter = 20000,
  ...
)
}
\arguments{
\item{x}{Object of class \code{icers}.}

\item{currency}{string. with currency used in the cost-effectiveness analysis (CEA).}

\item{effect_units}{string. unit of effectiveness}

\item{label}{whether to label strategies on the efficient frontier, all strategies, or none.
defaults to frontier.}

\item{label_max_char}{max number of characters to label the strategies - if not NULL (the default)
longer strategies are truncated to save space.}

\item{plot_frontier_only}{only plot the efficient frontier}

\item{alpha}{opacity of points}

\item{xlim, ylim}{Limits for the x and y axes. Text labels will be constrained
to these limits. By default, text labels are constrained to the entire plot
area.}

\item{max.iter}{Maximum number of iterations to try to resolve overlaps.
Defaults to 10000.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There are
  three types of arguments you can use here:

  \itemize{
    \item Aesthetics: to set an aesthetic to a fixed value, like
       \code{colour = "red"} or \code{size = 3}.
    \item Other arguments to the layer, for example you override the
      default \code{stat} associated with the layer.
    \item Other arguments passed on to the stat.
  }}
}
\value{
a ggplot2 object which can be modified by adding additional geoms
}
\description{
\code{plot.icers} plots the cost-effectiveness plane for a ICER object, calculated with \code{\link{calculate_icers}}
}
