% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_m_TR}
\alias{plot_m_TR}
\title{Plot cohort trace}
\usage{
plot_m_TR(m_M)
}
\arguments{
\item{m_M}{a cohort trace matrix}
}
\value{
a plot of the cohort trace
}
\description{
\code{plot_m_TR} plots cohort trace
}
