% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_m_TR_shiny}
\alias{plot_m_TR_shiny}
\title{Plot cohort trace for the Shiny App}
\usage{
plot_m_TR_shiny(m_M, input_list = NULL)
}
\arguments{
\item{m_M}{a cohort trace matrix}
}
\value{
a plot of the cohort trace for Shiny App
}
\description{
\code{plot_m_TR_shiny} plots cohort trace for the Shiny App
}
