% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Plot prevalence curve}
\usage{
plot_prevalence(l_m_M, v_names_sick_states, v_names_dead_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a ggplot object - plot of the prevalence curve
}
\description{
\code{plot_prevalence} plots the prevalence curve for specified health states.
}

