% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_proportion_sicker}
\alias{plot_proportion_sicker}
\title{Plot state-in-state proportion curve}
\usage{
plot_proportion_sicker(l_m_M, v_names_sick_states, v_names_sicker_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a ggplot object - plot of state-in-state proportion curve
}
\description{
\code{plot_prevalence} plots the
}

