% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_surv}
\alias{plot_surv}
\title{Plot survival curve}
\usage{
plot_surv(l_m_M, v_names_death_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a ggplot object - plot of the survival curve
}
\description{
\code{plot_surv} plots the survival probability curve.
}
