% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_tc}
\alias{plot_tc}
\title{Plot density of total cost}
\usage{
plot_tc(tc)
}
\arguments{
\item{tc}{total cost}
}
\value{
a plot of the density of total cost
}
\description{
\code{plot_tc} plots density of total cost.
}

