% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_te}
\alias{plot_te}
\title{Plot density of total QALYs}
\usage{
plot_te(te)
}
\arguments{
\item{tc}{total QALYs}
}
\value{
a plot of the density of total QALYs
}
\description{
\code{plot_te} plots density of total QALYs
}

