% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot cohort trace}
\usage{
plot_trace(m_M)
}
\arguments{
\item{m_M}{a cohort trace matrix}
}
\value{
a ggplot object - plot of the cohort trace
}
\description{
\code{plot_trace} plots the cohort trace.
}
