% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_trace_microsim}
\alias{plot_trace_microsim}
\title{Plot cohort trace of a microsimulation model}
\usage{
plot_trace_microsim(m_M)
}
\arguments{
\item{m_M}{a cohort trace matrix}
}
\value{
a plot of the cohort trace
}
\description{
\code{plot_trace_microsim} plots cohort trace of a microsimulation model.
}
