% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_trace_microsim_shiny}
\alias{plot_trace_microsim_shiny}
\title{Plot cohort trace of a microsimulation model for the Shiny App}
\usage{
plot_trace_microsim_shiny(m_M, input_list = NULL)
}
\arguments{
\item{m_M}{a cohort trace matrix}
}
\value{
a plot of the cohort trace for Shiny App
}
\description{
\code{plot_trace_microsim_shiny} plots cohort trace of a microsimulatoin model for the Shiny App.
}

