% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{plot_trace_PSM}
\alias{plot_trace_PSM}
\title{Plot Markov trace from a partitioned survival model.}
\usage{
plot_trace_PSM(time, partsurv.model, PA = F)
}
\arguments{
\item{time}{numeric vector of time to estimate probabilities.}

\item{partsurv.model}{partitioned survival model.}

\item{PA}{run probabilistic analysis.
Default = FALSE.}
}
\value{
a plot of the cohort trace.
}
\description{
\code{plot_trace_PSM} plots Markov trace from a partitioned survival model.
}
