% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_trace_strategy}
\alias{plot_trace_strategy}
\title{Plot cohort trace per strategy}
\usage{
plot_trace_strategy(l_m_M)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a ggplot object - plot of the cohort trace for each strategy split by health state.
}
\description{
\code{plot_trace} plots the cohort trace for each strategy, split by health state.
}
