% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_conversion.R
\name{prob_to_odds}
\alias{prob_to_odds}
\title{Convert a probability to an odds}
\usage{
prob_to_odds(p)
}
\arguments{
\item{p}{a scalar of vector of probabilities}
}
\value{
a scalar or vector of odds
}
\description{
\code{prob_to_odds} convert a probability to an odds.
}

