% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_conversion.R
\name{prob_to_prob}
\alias{prob_to_prob}
\title{Convert a probability to a probability with a different frequency}
\usage{
prob_to_prob(p, t = 1)
}
\arguments{
\item{p}{probability}

\item{t}{time/ frequency}
}
\value{
a number - converted probability
}
\description{
\code{rate_to_prob} convert a probability to a probability with a different frequency.
}
