% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_conversion.R
\name{rate_to_prob_old}
\alias{rate_to_prob_old}
\title{Convert a rate to a probability}
\usage{
rate_to_prob_old(r, t = 1)
}
\arguments{
\item{r}{rate}

\item{t}{number of cycles per base time unit (frequency)}
}
\value{
a scalar or vector with probabilities
}
\description{
\code{rate_to_prob_old} convert a rate to a probability.
}
\examples{
# Annual rate to monthly probability
r_year  <- 0.3
p_month <- rate_to_prob_old(r = r_year, t = 12)
p_month
}
