% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{runMLE.cure}
\alias{runMLE.cure}
\title{run MLE on cure models}
\usage{
runMLE.cure(x, exArgs)
}
\arguments{
\item{x}{Data or model object used to estimate the cure model by MLE.}

\item{exArgs}{Named list of additional arguments passed to the optimizer/fitting routine.}
}
\description{
run MLE on cure models
}
