% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplev.R
\name{samplev}
\alias{samplev}
\title{Vectorized categorical distribution}
\usage{
samplev(m.Probs)
}
\arguments{
\item{p}{probability}

\item{t}{time/ frequency}
}
\value{
a number - converted rate
}
\description{
\code{samplev} sample states for multiple individuals simultaneously.
}
