% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{surv_to_haz}
\alias{surv_to_haz}
\title{Convert survival probabilities to hazard rates.}
\usage{
surv_to_haz(surv, times)
}
\arguments{
\item{surv}{vector of survival probabilities.}

\item{times}{vector of time}
}
\value{
vector of hazard rates
}
\description{
\code{surv_to_haz} convert survival probabilities to hazard rates.
}
