% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{trace.DES}
\alias{trace.DES}
\title{Compute Markov trace out of a multi-state model using DES.}
\usage{
trace.DES(msm_sim = des_sim, tmat, n_i, times)
}
\arguments{
\item{msm_sim}{multi-state model.}

\item{tmat}{matrix of transition history.}

\item{n_i}{number of individuals.}

\item{times}{time horizon the extrapolation is done over.}
}
\value{
Matrix of Markov trace
}
\description{
\code{trace.DES} computes Markov trace out of a multi-state model using DES.
}
