% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{trans_prob}
\alias{trans_prob}
\title{Calculate transition probabilities.}
\usage{
trans_prob(surv)
}
\arguments{
\item{surv}{vector of survival probabilities.}
}
\value{
matrix of transition probabilities
}
\description{
\code{trans_prob} calculates transition probabilities using survival probabilities.
}
