% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{trans_to_surv}
\alias{trans_to_surv}
\title{Convert transition probabilities back to survival probabilities.}
\usage{
trans_to_surv(t.p)
}
\arguments{
\item{t.p}{vector of transition probabilities.}
}
\value{
vector of survival probabilities
}
\description{
\code{trans_to_surv} convert transition probabilities back to survival probabilities.
}
