% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calc_prop_sicker}
\alias{calc_prop_sicker}
\title{Calculate state-in-state proportions}
\usage{
calc_prop_sicker(l_m_M, v_names_sick_states, v_names_sicker_states)
}
\arguments{
\item{l_m_M}{a list containing cohort trace matrices}
}
\value{
a dataframe containing state-in-state proportions of specified health states for each strategy
}
\description{
\code{plot_prevalence} calculates the proportion of a speciefied subset of states among a set of specified states
}
