% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fit.fun.cure}
\alias{fit.fun.cure}
\title{Fit multiple mixture cure models on survival data}
\usage{
fit.fun.cure(
  time,
  status,
  data = data,
  extrapolate = FALSE,
  times,
  legend_position = "top",
  xlow = min(times),
  xhigh = max(times),
  ylow = 0,
  yhigh = 1,
  risktable = F,
  mods = c("exp", "weibull", "gamma", "lnorm", "llogis", "gompertz", "rps")
)
}
\arguments{
\item{time}{numeric vector of time to estimate probabilities.}

\item{status}{numeric vector of event status.}

\item{data}{dataframe containing the time and status variables.}

\item{extrapolate}{extrapolate beyond model time horizon.
Default = FALSE.}

\item{times}{time horizon the extrapolation is done over.}

\item{legend_position}{position of the legend.
Default = "top".}

\item{xlow}{time horizon the extrapolation is done over.
Default = min(time).}

\item{xhigh}{time horizon the extrapolation is done over.
Default = max(time).}

\item{ylow}{time horizon the extrapolation is done over.
Default = 0.}

\item{yhigh}{time horizon the extrapolation is done over.
Default = 1.}

\item{risktable}{time horizon the extrapolation is done over.
Default = F.}

\item{mods}{a vector of models to fit.
Default = c("exp", "weibull", "gamma", "lnorm", "llogis", "gompertz", "rps").}
}
\value{
a list containing all survival model objects.
}
\description{
\code{fit.fun.cure} fits multiple mixture cure models to survival data using flexsurv.
}
