% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fit.fun}
\alias{fit.fun}
\title{Fit multiple survival models on survival data}
\usage{
fit.fun(
  time,
  status,
  data = data,
  extrapolate = FALSE,
  times,
  k = 2,
  legend_position = "top",
  xlow = min(times),
  xhigh = max(times),
  ylow = 0,
  yhigh = 1,
  risktable = F
)
}
\arguments{
\item{time}{numeric vector of time to estimate probabilities.}

\item{status}{numeric vector of event status.}

\item{data}{dataframe containing the time and status variables.}

\item{extrapolate}{extrapolate beyond model time horizon.
Default = FALSE.}

\item{times}{time horizon the extrapolation is done over.}

\item{k}{number of knots in Royston-Parmar spline model.
Default = 2.}

\item{legend_position}{position of the legend.
Default = "top".}

\item{xlow}{time horizon the extrapolation is done over.
Default = min(time).}

\item{xhigh}{time horizon the extrapolation is done over.
Default = max(time).}

\item{ylow}{time horizon the extrapolation is done over.
Default = 0.}

\item{yhigh}{time horizon the extrapolation is done over.
Default = 1.}

\item{risktable}{time horizon the extrapolation is done over.
Default = F.}
}
\value{
a list containing all survival model objects.
}
\description{
\code{fit.fun} fits multiple survival models to survival data using survHE.
}
