% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{plot_trace_microsim}
\alias{plot_trace_microsim}
\title{Plot cohort trace from microsimulation model}
\usage{
plot_trace_microsim(m_M)
}
\arguments{
\item{m_M}{cohort trace array from microsimulation model.}
}
\value{
Plot of the cohort trace.
}
\description{
\code{plot_trace_microsim} computes Markov trace out of a multi-state model using DES.
}
