% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_conversion.R
\name{rate_to_prob}
\alias{rate_to_prob}
\title{Convert a rate to a probability}
\usage{
rate_to_prob(r, t = 1)
}
\arguments{
\item{r}{rate}

\item{t}{time/ frequency}
}
\value{
a scalar or vector with probabilities
}
\description{
\code{rate_to_prob} convert a rate to a probability.
}
\examples{
# Annual rate to monthly probability
r_year  <- 0.3
r_month <- rate_to_prob(r = r_year, t = 1/12)
r_month
}

