% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplev.R
\name{samplev}
\alias{samplev}
\title{Vectorized categorical distribution}
\usage{
samplev(m_Probs, m = 1)
}
\arguments{
\item{m_Probs}{matrix with probabilities for n_i individual and n_states
states}

\item{m}{number of time cycles to sample}
}
\value{
v_cat: An n_i x 1 matrix filled with sampled health state(s) per
individual
}
\description{
\code{samplev} sample states for multiple individuals simultaneously.
}

