% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{surv_prob}
\alias{surv_prob}
\title{Calculate survival probabilities.}
\usage{
surv_prob(model, times = NULL, PA = FALSE, rx = 1)
}
\arguments{
\item{model}{survival model.}

\item{times}{time horizon to calculate survival probabilities.}

\item{PA}{run probabilistic analysis.}

\item{rx}{determines which treatment arm (same order as factor levels of treatment variable).
Default = FALSE.}
}
\value{
vector of survival probabilities.
}
\description{
\code{surv_prob} calculates survival probabilities from survival models.
}
